Implemented Race Class
Successfully implemented the first version of the Race class. Properties for Info and Performance instances have been commented out due to a lack of implementation of these classes at present.

Need to adjust logic of constructor such that the user can specify whether or not to include exhibition races as part of the "round" index parameter when searching.